requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djoser==1.3.1
psycopg2==2.7.6.1
PyJWT==1.6.4
pytz==2018.7
